+++ src/components/OverviewItem/stories/OverviewItemDepiction.stories.tsx
@@ -3,7 +3,7 @@ import { Meta, StoryFn } from "@storybook/react";
 
 import png16to9 from "../../Depiction/stories/test-16to9.png";
 
-import { Depiction, Icon, OverviewItem, OverviewItemDepiction } from "./../../../../index";
+import { Depiction, Icon, OverviewItem, OverviewItemDepiction } from "./../../../components";
 import { FullExample as DepictionExample } from "./../../Depiction/stories/Depiction.stories";
 
 export default {